backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
requests
fpdf2
reportlab
langchain
langchain-core
langchain-openai
langchain-google-genai
langchain-community
langchain-tavily
langgraph-checkpoint-postgres
psycopg[binary,pool]
pdfplumber
python-docx
openpyxl
python-pptx
aiosqlite